Started implementing reflective database interface
Everything is working so far- except SQLAlchemy primary key generation
is a big FAIL.  Unless the primary key is explicitly created by the
programmer, SQLAlchemy will issue a "null value for non-nullable entry"
error when it should be auto-generating the key by it's self.

Will attempt to isolate a test case and post to Stack Overflow and other
lists.

If there is no resolution, will abandon reflective database interface
and switch to declarative.

The disadvantage of a declarative interface is that it requires
duplicating and maintaining aditional parts of the DB schema-
specifically Foreign Key relations and CASCADE switches.