Refactored to a declarative-reflective style
Thanks to an example in the GeoAlchemy source, I became aware of a way
to fuse the declarative and reflexive styles for declaring tables and
linking objects with SQLAlchemy.  The source code has been refactored to
take advantage of this as it eliminates a lot of repetition and some
kludgey statements.