Fixed mandatory arguments to class constructors
The database classes had mandatory arguments for each attribute.  I.E.
when creating a new source you had to specify a value for EVERY column
in the database table.

Each argument to the class constructors now defaults to None.  If an
argument is required, the NOT NULL specification in the database schema
will catch it.

Also fixed some bugs with variable names in class constructors.