Removed getter/setter methods in DB classes- switched to synonyms
I was adding a lot of getter/setter methods to the Base class in DBman
which provided each object generated from database records the following
functionality:

   windRecord.getDateTime() -> returns windRecord.windatetime

Or for a wave record:

   waveRecord.getDateTime() -> returns waveRecord.wavdatetime

This was implementing by grepping through the object's attribute list
looking for something that ended in "datetime".  It was getting out of
control and utterly failed to work for things like IDs where there are
multiple columns with the words ID in them.

Threw this approach out and switched to SQLAlchemy synonyms which allow
this functionality to be implemented cleanly on a case-by-case basis in
the class definitions.

Currently only the Source and Wind classes have this implemented.  More
will follow.