Started top-level documentation
Each WaveConnect component has stand-alone documentation, which is very nice for
shipping deliverables that contain only one sub-folder of the whole project.

However, future devliverables will likely ship more than one component,
therefore it is necessary to provide a way of integrating the Sphinx
documentation while at the same time retaining the independence of the current
system.

This is done by creating a new Sphinx project in the manual folder of the
top-level doc folder.  The source directory of this project contains symlinks to
the source folders of sub-project documentation.

This approach works very well- the only problem is that Windows has no concept
of what a symlink is--- on those systems git will just place a file containing
the path that was symlinked.  I.E:

    PyModules -> ../../../py/doc/source

Will just be a text file containing:

    ../../../py/doc/source

And not a functioning link.

Can't think of a better way to do it at this time, but the system is simple
enough that it could be replaced if a workaround is found for this issue.