Wave records now committed to database.
This took way. longer. than. it. should. have.

The majority of time wasted on this was due to figuring out how to select wave
spectra records:

    .filter( SpectraRecord.spectrafreq == somePythonList )

Problem was SQLAlchemy was rendering `somePythonList` as an ARRAY of PostgreSQL
NUMERIC entities while `spectrafreq` was an array of double precision values.
Hence, the equality comparison failed with a type error.

The solution is a cocktail of the SQLAlchemy `cast()` function along with some
PostgreSQL datatypes:

    from sqlalchemy.sql.expression import cast
    from sqlalchemy.dialects.postgresql import ARRAY, DOUBLE_PRECISION

Then, do the following:

    .filter( SpectraRecord.spectrafreq == cast(
      somePythonList, ARRAY(DOUBLE_PRECISION)
    ))

Hopefully I won't have to waste 2 hours on that ever again.