Implemented disaggregation for wave spectra
If all direction spectra are present, a 2D spectra can now be produced using the
formulas supplied by the NDBC.  The number of bins used can be specified using
the `--num-dir-bins` flag to getNDBCData.py.

If 0 bins are specified and all the direction spectra are present, the spectra
are collapsed into a 5 row matrix in the order of:

- Row 1: density
- Row 2: Alpha 1
- Row 3: Alpha 2
- Row 4: R1
- Row 5: R2

If only the density spectra is present, it is saved as a 1D array.

The end result is that all spectra in a wave record should be replaced with one
'spectra' entry and a 'spectraType' entry which is:

- '1D': A vector containing the density spectra.
- '2D': A matrix containing the disaggregated density and direction spectra.
- '2D-aggregated': A matrix containing the density and direction spectra.