Added Pip and associated resources
The pip instalation specification is quite hacky---it relies on
`easy_install-2.7` being present in the path.

The rest of the files add support so that Pip may be used by Chef as a provider
of Python packages.  The provider and resource files are taken from Christopher
Peplin's repository as discussed in Chef issue 973:

    http://tickets.opscode.com/browse/CHEF-973

It looks like a native pip provider may show up in the 0.10 release of Chef.

Modifications had to be made in order for the pip provider to be able to inherit
from the EasyInstall class outside of the Chef install location.  This was
acheieved with some tricks borrowed from Corey Donohoe's Smeagol repository
which pulls similar tricks for Homebrew on OS X:

    https://github.com/atmos/smeagol