Finished cmcards parser
The parser is now constrained to being line-oriented and no longer allows
matches to spread accross line boundaries.  This probably matchs the way the
Fortran routine in CMS reads the file---one line at a time.

The parser seems to process the cmcards file for the Humboldt bay simulation
just fine---however there could be lurking bungs.  All in all a declarative
parser that fits in ~30 lines is way better than 50 lines of regular expressions
that couldn't even process the whole file.