Created new IO module from NDBC JSON and matlab modules
The CMS post-processor will support these output formats so they are not
specific to NDBC anymore.  A new IO (input/output) module has been created so
that they may be shared among different tools.

Database output routines are still specialized to a degree that they cannot be
generalized---although this would be awesome.

Fixed a couple of buglets in the NDBC downloader.py.