Added parser for CMS-Flow telescoping grid files
The parser extracts columns of x coordinates and y coordinates from a CMS .tel
file and returns them as an Numpy array.

Also included is a function that rotates the computational grid, shifts it to
the orgin coordinates and then transforms the resulting locations to WGS84.
Adding this functionality required the use of the pyproj module which is an
interface to the PROJ4 library.