Current records are now extracted
The CMS postprocess is now capable of parsing and extracting CMS-Flow current
information, pairing it with location and timestep information and storing the
results as a list of dicts.  Committing the dicts to the database is a trivial
next step.

However, for a 24 hour run of the humboldt domain with hourly output the
postprocessor takes ~half a minute to create the dictionaries and dump them to
JSON.  Memory usage is around 1 GB.  For writing to matlab files, memory usage
increases to 1.5 GB and execution time is... still going.

Could try switching from list comprehensions to generator comprehensions so that
the entire dataset is never explicitly created in-memory---but this has serious
potential to get out of hand for larger runs.