Silly rabbit, (x, y) is (lon, lat) not (lat, lon)
This places the grid in the right place---however the rotation is off.  It could
be that the grid angle is given in compass degrees.

Before sorting out the angle, I need to deal with the amount of time it takes to
transfer information into the database.  30+ minutes and 10 GB of RAM is just
not going to work especially when the dataset starts increasing in size.