Improved efficiency of posprocessing
CMS datasets are large.  This change uses generator comprehensions and itertools
components to relcace the list comprehension that creates the current record
dictionaries.  The resulting generator comprehension is more space efficient and
allows records to be dealt with one at a time or in small batches in order to
save on RAM.

JSON output routines in the IO module have been modified to deal with generator
objects.