Efficient import of large datatsets to Postgres
Added a new routine, `bulk_import`, to DBman. `bulk_import` only works with
Postgres databases using the `psycopg2` Python module and takes advantage of the
efficiency of the COPY statement.  `bulk_import` takes a list of Python
dictionaries containing key/value pairs that match the column names and
datatypes of the table targetted for import.  A generator that can produce such
a list can also be provided and is prefered if a large amount of data is to be
imported.

`bulk_import` works by generating a temporary CSV file from the dictionaries and
then using COPY to import this file into a temporary Postgres table.  The
contents of the temporary table are then inserted into the target table and the
temporary table is dropped.  The COPY process takes about a fourth of the time that
batched INSERT statements would require.

The speed of `bulk_import` could be improved ~33% by cutting the temporary CSV
file out of the process---but this would require some clever programming.