Cleaned up current record generator
The generator in postprocessing.py was originally a generator comprehension that
evolved from a list comprehension.  The list comprehension was a little
indecipherable, but created records in 30 seconds compared to 1 minute 30
seconds for a pair of nested loops.

This monstrosity has been replaced with a generator constructed from a pair of
nested loops returning values using `yield`.  It is much easier to decipher.